Create separate routes file
Requires the app to be defined
in the init file, and for the
dockerfile to install falcon as
a package for the imports to work
