+++ Dockerfile
@@ -3,10 +3,10 @@ FROM python:3.6
 RUN mkdir /falcon
 WORKDIR /falcon
 
-COPY requirements.txt .
-
-RUN pip install -r requirements.txt
-
 COPY . .
+
+RUN pip install -U setuptools
+
+RUN pip install .
 
 CMD ["python", "./falcon/__main__.py"]
