+++ falcon/__main__.py
@@ -1,24 +1,8 @@
 import os
 import threading
-from flask import Flask
-from igniter import Igniter
-from queue_handler import QueueHandler
-
-
-FALCON_THREAD_NAMES = ['queueHandler', 'igniter']
-
-app = Flask(__name__)
-
-
-@app.route("/health")
-def status():
-    active_threads = threading.enumerate()
-    active_falcon_threads = [thread for thread in active_threads if thread.name in FALCON_THREAD_NAMES]
-    for thread in active_falcon_threads:
-        if not thread.is_alive():
-            return "Error in {} thread {}".format(thread.name, thread.ident)
-    healthy_threads = ['{}-{}'.format(t.name, t.ident) for t in active_falcon_threads]
-    return 'Falcon threads: ' + ', '.join(healthy_threads)
+from falcon import app
+from falcon.igniter import Igniter
+from falcon.queue_handler import QueueHandler
 
 
 class WebThread(threading.Thread):
@@ -28,7 +12,6 @@ class WebThread(threading.Thread):
 
 if __name__ == '__main__':
     config_path = os.environ.get('CONFIG_PATH')
-
     handler = QueueHandler(config_path)  # instantiate a concrete `QueueHandler`
     igniter = Igniter(config_path)  # instantiate a concrete `Igniter`
     w = WebThread()
