+++ falcon/igniter.py
@@ -7,8 +7,8 @@ from threading import Thread, get_ident
 import requests
 from cromwell_tools.cromwell_api import CromwellAPI
 
-import queue_handler
-import settings
+from falcon import queue_handler
+from falcon import settings
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger('falcon.{module_path}'.format(module_path=__name__))
